use appropriate length variables to save code space
git-svn-id: svn+ssh://svn.code.sf.net/p/openlcb/svn/trunk/prototypes/C@1842 87536cb5-86f1-4cdc-8fc6-3f0f71bcd267

+++ libraries/OlcbArduinoCAN/BG.cpp
@@ -8,7 +8,7 @@
 #define UNREADY_BLINK 0xFF00FFL
 #define READY_BLINK   0x1L
 
-BG::BG(PCE* pc, ButtonLed** bC, long* pt, int n, ButtonLed* bptr, ButtonLed* gptr) {
+BG::BG(PCE* pc, ButtonLed** bC, long* pt, uint8_t n, ButtonLed* bptr, ButtonLed* gptr) {
       pce = pc;
       buttons = bC;
       patterns = pt;
